+++ tests/testthat/test-Rbuildignore.R
@@ -10,3 +10,11 @@ test_that("roxygen ignores files with matching pattern in .Rbuildignore", {
   writeChar("^R/ignore_me.R$\n", file.path(test_pkg, ".Rbuildignore"), eos = NULL)
   expect_equal(basename(package_files(test_pkg)), "a.R")
 })
+
+test_that("roxygen works with empty lines in .Rbuildignore", {
+  test_pkg <- temp_copy_pkg(test_path("testRbuildignore"))
+  on.exit(unlink(test_pkg, recursive = TRUE))
+
+  writeChar("^R/ignore_me.R$\n\n.nonexistentfile", file.path(test_pkg, ".Rbuildignore"), eos = NULL)
+  expect_equal(basename(package_files(test_pkg)), "a.R")
+})